requirements.txt
python-dotenv
langchain
langchain-community
langchain-core
streamlit
unstructured[all-docs]
chromadb==0.3.29
sentence-transformers
torch
transformers
accelerate
llama-cpp-python
python-docx
pytesseract
pdf2image
Pillow
tqdm
pypdf
docx2txt
